made Auction Bonus's be on 1 line
also started on undo (current plan is only allowing undo of trades in current state (for current player, obviously).  Not going to allow any other undo at this time. either have to queue the trades, and not do them until state transition. or do it on server side.  not sure yet